backend/app/agents/code_review/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
GitPython==3.1.40
PyGithub==1.59.1
google-generativeai==0.8.0
requests==2.31.0
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
openai==1.3.0